Removing class introduced as part of a death path
