+++ core/persistence-jpa/src/main/java/org/apache/syncope/core/persistence/jpa/entity/task/JPASchedTask.java
@@ -18,6 +18,7 @@
  */
 package org.apache.syncope.core.persistence.jpa.entity.task;
 
+import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
@@ -25,20 +26,40 @@ import javax.persistence.Entity;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
 import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+import org.apache.syncope.common.lib.types.IdRepoImplementationType;
+import org.apache.syncope.core.persistence.api.entity.Implementation;
 import org.apache.syncope.core.persistence.api.entity.task.SchedTask;
 import org.apache.syncope.core.persistence.api.entity.task.TaskExec;
+import org.apache.syncope.core.persistence.jpa.entity.JPAImplementation;
 import org.apache.syncope.core.persistence.jpa.validation.entity.SchedTaskCheck;
 
 @Entity
 @Table(name = JPASchedTask.TABLE)
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 @SchedTaskCheck
-public class JPASchedTask extends AbstractSchedTask<SchedTask> implements SchedTask {
+public class JPASchedTask extends AbstractTask<SchedTask> implements SchedTask {
 
     private static final long serialVersionUID = 7596236684832602180L;
 
     public static final String TABLE = "SchedTask";
+
+    @OneToOne(optional = false)
+    private JPAImplementation jobDelegate;
+
+    private OffsetDateTime startAt;
+
+    private String cronExpression;
+
+    @NotNull
+    private String name;
+
+    private String description;
+
+    @NotNull
+    private Boolean active = true;
 
     @OneToMany(targetEntity = JPASchedTaskExec.class,
             cascade = CascadeType.ALL, orphanRemoval = true, mappedBy = "task")
@@ -52,5 +73,67 @@ public class JPASchedTask extends AbstractSchedTask<SchedTask> implements SchedT
     @Override
     protected List<TaskExec<SchedTask>> executions() {
         return executions;
+    }
+
+    @Override
+    public Implementation getJobDelegate() {
+        return jobDelegate;
+    }
+
+    @Override
+    public void setJobDelegate(final Implementation jobDelegate) {
+        checkType(jobDelegate, JPAImplementation.class);
+        checkImplementationType(jobDelegate, IdRepoImplementationType.TASKJOB_DELEGATE);
+        this.jobDelegate = (JPAImplementation) jobDelegate;
+    }
+
+    @Override
+    public OffsetDateTime getStartAt() {
+        return startAt;
+    }
+
+    @Override
+    public void setStartAt(final OffsetDateTime startAt) {
+        this.startAt = startAt;
+    }
+
+    @Override
+    public String getCronExpression() {
+        return cronExpression;
+    }
+
+    @Override
+    public void setCronExpression(final String cronExpression) {
+        this.cronExpression = cronExpression;
+    }
+
+    @Override
+    public String getDescription() {
+        return description;
+    }
+
+    @Override
+    public void setDescription(final String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public void setName(final String name) {
+        this.name = name;
+    }
+
+    @Override
+    public boolean isActive() {
+        return active;
+    }
+
+    @Override
+    public void setActive(final boolean active) {
+        this.active = active;
     }
 }
